backend-agent/requirements.txt
fastapi
uvicorn
langchain-google-vertexai
langchain-postgres
langchain-google-firestore
google-cloud-dlp
pgvector
asyncpg
psycopg[binary]
pydantic-settings
pypdf
langchain-community
google-auth
requests
slowapi
redis
langchain-redis
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-gcp-trace
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
firebase-admin
tenacity
opentelemetry-instrumentation-langchain
stripe
sqlalchemy
google-cloud-secret-manager